requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.6
django-cors-headers==3.13.0
django-email-verification==0.2.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-excel==2.1.0
et-xmlfile==1.1.0
gunicorn==20.1.0
numpy==1.23.1
openpyxl==3.0.10
pandas==1.4.3
psycopg2==2.9.3
PyJWT==2.4.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
